pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "plaid-cli"
version = "0.1.0"
description = "CLI tool to pull personal bank transactions via Plaid API"
requires-python = ">=3.10"
dependencies = [
    "plaid-python>=29.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
plaid = "plaid_cli.router:main"

[tool.setuptools.packages.find]
include = ["plaid_cli*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
